#537 Separated table component
